src/requirements.txt
certifi==2022.12.7
databases==0.5.3
fastapi==0.92.0
httpx==0.24.0
prometheus-client==0.16.0
prometheus-fastapi-instrumentator==6.0.0
pydantic==1.8.2
python-dotenv==0.21.0
requests==2.26.0
requests-toolbelt==1.0.0
SQLAlchemy==1.4.26
starlette==0.25.0
urllib3==1.26.7
uvicorn==0.15.0
pytz
psycopg2-binary
minio
loguru
python-dotenv==0.21.0
pandas